+++ src/backend/cpu/CpuWorker.cpp
@@ -43,9 +43,8 @@
 
 #ifdef XMRIG_ALGO_RANDOMX
 #   include "crypto/randomx/randomx.h"
+#   include "crypto/randomx/sha256/sha256.h"
 #endif
-
-
 
 
 namespace xmrig {
@@ -306,6 +305,10 @@ void xmrig::CpuWorker<N>::start()
                     job.generateMinerSignature(m_job.blob(), job.size(), miner_signature_ptr);
                 }
                 randomx_calculate_hash_next(m_vm, tempHash, m_job.blob(), job.size(), m_hash);
+
+                if (job.algorithm() == Algorithm::RX_TUSKE) {
+                    SHA256d_Buf(m_hash, RANDOMX_HASH_SIZE, m_hash);
+                }
             }
             else
 #           endif