+++ src/base/crypto/Algorithm.cpp
@@ -88,6 +88,7 @@ const char *Algorithm::kRX_SFX          = "rx/sfx";
 const char *Algorithm::kRX_KEVA         = "rx/keva";
 const char *Algorithm::kRX_YADA         = "rx/yada";
 const char *Algorithm::kRX_XDAG         = "rx/xdag";
+const char *Algorithm::kRX_TUSKE        = "rx/tuske";
 #endif
 
 #ifdef XMRIG_ALGO_ARGON2
@@ -161,6 +162,7 @@ static const std::map<uint32_t, const char *> kAlgorithmNames = {
     ALGO_NAME(RX_KEVA),
     ALGO_NAME(RX_YADA),
     ALGO_NAME(RX_XDAG),
+    ALGO_NAME(RX_TUSKE),
 #   endif
 
 #   ifdef XMRIG_ALGO_ARGON2
@@ -288,6 +290,8 @@ static const std::map<const char *, Algorithm::Id, aliasCompare> kAlgorithmAlias
                                     ALGO_ALIAS(RX_YADA,         "randomyada"),
     ALGO_ALIAS_AUTO(RX_XDAG),       ALGO_ALIAS(RX_XDAG,         "randomx/xdag"),
                                     ALGO_ALIAS(RX_XDAG,         "randomxdag"),
+    ALGO_ALIAS_AUTO(RX_TUSKE),      ALGO_ALIAS(RX_TUSKE,        "randomx/tuske"),
+                                    ALGO_ALIAS(RX_TUSKE,        "randomtuske"),
 #   endif
 
 #   ifdef XMRIG_ALGO_ARGON2
@@ -379,7 +383,7 @@ std::vector<xmrig::Algorithm> xmrig::Algorithm::all(const std::function<bool(con
         CN_PICO_0, CN_PICO_TLO,
         CN_UPX2,
         CN_GPU,
-        RX_0, RX_WOW, RX_ARQ, RX_GRAFT, RX_SFX, RX_KEVA, RX_YADA, RX_XDAG,
+        RX_0, RX_WOW, RX_ARQ, RX_GRAFT, RX_SFX, RX_KEVA, RX_YADA, RX_XDAG, RX_TUSKE,
         AR2_CHUKWA, AR2_CHUKWA_V2, AR2_WRKZ,
         KAWPOW_RVN,
         GHOSTRIDER_RTM, GHOSTRIDER_MIKE