+++ src/crypto/rx/RxAlgo.cpp
@@ -50,6 +50,9 @@ const RandomX_ConfigurationBase *xmrig::RxAlgo::base(Algorithm::Id algorithm)
     case Algorithm::RX_YADA:
         return &RandomX_YadaConfig;
 
+    case Algorithm::RX_TUSKE:
+        return &RandomX_TuskeConfig;
+
     default:
         break;
     }